token, login and authorization

+++ osa4/blogilista/controllers/blogs.js
@@ -1,5 +1,7 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 const getTokenFrom = request => {
   const authorization = request.get('authorization')
@@ -17,7 +19,13 @@ blogRouter.get('/', (request, response) => {
     })
 })
 
-blogRouter.post('/', (request, response) => {
+blogRouter.post('/', async (request, response) => {
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   if (!request.body.title) {
     return response.status(400).json({
       error: 'Title is missing'
@@ -34,16 +42,19 @@ blogRouter.post('/', (request, response) => {
       "title": request.body.title,
       "author": request.body.author,
       "url": request.body.url,
+      "user": {
+        "username": user.username,
+        "name": user.name,
+        "id": user._id
+      },
       "likes": request.body.likes | 0
     })
 
-  blog
-    .save()
-    .then(savedBlog => savedBlog.toJSON())
-    .then(savedAndFormattedBlog => {
-      response.json(savedAndFormattedBlog)
-    })
-    .catch(error => next(error))
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog.toJSON())
 })
 
 blogRouter.put('/:id', (request, response, next) => {
@@ -63,13 +74,20 @@ blogRouter.put('/:id', (request, response, next) => {
     .catch(error => next(error))
 })
 
-blogRouter.delete('/:id', (request, response, next) => {
+blogRouter.delete('/:id',async (request, response, next) => {
   const id = request.params.id
-  Blog.findByIdAndRemove(id)
-    .then(result => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(id)
+  if (user.username===blog.user.username) {
+    const deletedBlog = await Blog.findById(id)
+    user.blogs = user.blogs.splice(deletedBlog._id)
+    await user.save()
+    await Blog.findByIdAndRemove(id)
+  }
 })
 
 module.exports = blogRouter
